# dbms/static/swagger.yaml
openapi: 3.0.3
info:
    title: DBMS Recipe API
    description: RESTful API for managing users, recipes, ingredients, and nutrition.
    version: 1.0.0
servers:
    - url: /
      description: Local Server
components:
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            in: header
            name: dbms-api-key
            description: Provide your API key to access protected endpoints.
    schemas:
        User:
            type: object
            properties:
                id:
                    type: integer
                username:
                    type: string
                email:
                    type: string
                created_at:
                    type: string
                    format: date-time
                allergies:
                    type: string
                    nullable: true
        Recipe:
            type: object
            properties:
                id:
                    type: integer
                title:
                    type: string
                procedure:
                    type: string
                    nullable: true
                created_at:
                    type: string
                    format: date-time
                servings:
                    type: integer
                    nullable: true
                cuisine_type:
                    type: string
                    nullable: true
                cooking_methods:
                    type: string
                    nullable: true
                created_by:
                    type: integer
        Ingredient:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                img_url:
                    type: string
                    nullable: true
                allergy:
                    type: string
                    nullable: true
                calories:
                    type: number
                carbs:
                    type: number
                protein:
                    type: number
                fat:
                    type: number
...
paths:
...
    /api/users/:
        get:
            tags: [Users]
            summary: Get a paginated list of users
            parameters:
                - name: limit
                  in: query
                  schema:
                      type: integer
                      default: 10
                - name: offset
                  in: query
                  schema:
                      type: integer
                      default: 0
            responses:
                "200":
                    description: A list of users
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/User"
        post:
            tags: [Users]
            summary: Register a new user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [username, email, pwd]
                            properties:
                                username:
                                    type: string
                                email:
                                    type: string
                                pwd:
                                    type: string
            responses:
                "201":
                    description: User created successfully
                    headers:
                        Location:
                            schema:
                                type: string
                "400":
                    description: Bad request (validation error)
                "409":
                    description: Username or email already exists
...
    /api/recipes/:
        get:
            tags: [Recipes]
            summary: Get paginated recipes
            parameters:
                - name: limit
                  in: query
                  schema:
                      type: integer
                      default: 10
                - name: offset
                  in: query
                  schema:
                      type: integer
                      default: 0
            responses:
                "200":
                    description: List of recipes
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Recipe"
        post:
            tags: [Recipes]
            summary: Create a new recipe
            security:
                - ApiKeyAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [title]
                            properties:
                                title:
                                    type: string
                                procedure:
                                    type: string
                                servings:
                                    type: integer
                                cuisine_type:
                                    type: string
            responses:
                "201":
                    description: Recipe created
...
    /api/ingredients/:
        get:
            tags: [Ingredients]
            summary: Get paginated ingredients
            parameters:
                - name: limit
                  in: query
                  schema:
                      type: integer
                      default: 10
                - name: offset
                  in: query
                  schema:
                      type: integer
                      default: 0
            responses:
                "200":
                    description: List of ingredients
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Ingredient"
        post:
            tags: [Ingredients]
            summary: Add a new ingredient to database
            security:
                - ApiKeyAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [name]
                            properties:
                                name:
                                    type: string
                                calories:
                                    type: number
                                protein:
                                    type: number
                                carbs:
                                    type: number
                                fat:
                                    type: number
            responses:
                "201":
                    description: Ingredient created
...